Refactor: libcrmcommon: New pcmk__compare_versions()
To replace compare_version()

The next commit replaces this with a regex implementation that I think I
like better. We can squash the commits if we decide to use that.

We could also take a middle ground approach of splitting each version
string on dots using g_strsplit() and then parsing each segment that
way, without involving a regex. All we'd have to do is iterate over each
segment to make sure it's only digits. Not sure which one is more
readable. Probably the regex approach, but regex.h functions are pretty
ugly.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>