Refactor: libcrmcommon, executor: Drop space before variadic arg comma
See the removed comment for more info on why these gratuitous spaces
were present in logging.h and cts-exec-helper.c.

We don't use the extra space in other files that use '##args'
(scheduler_internal.h, cib_utils.c, and output_log.c). And gcc fixed
that bug in version 3.0.0, which was released in 2001. We should be
okay.

Details:
https://gcc.gnu.org/onlinedocs/gcc-4.2.2/cpp/Differences-from-previous-versions.html#Differences-from-previous-versions
https://gcc.gnu.org/gcc-3.0/

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>