Refactor: libcrmcommon: New PCMK__LOG_STDOUT
To replace LOG_STDOUT.

We can't even deprecate LOG_STDOUT until the public macros that use it
are deprecated.

The goal here is to get internal constants for LOG_STDOUT, LOG_NEVER,
and ultimately LOG_TRACE. The prefix will be rather cumbersome for
PCMK__LOG_TRACE. However, I want to make it clear that these log levels
are custom extensions and are not broadly supported symbols.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>