Refactor: libcrmcommon: New PCMK__LOG_TRACE
To replace LOG_TRACE.

We can't even deprecate LOG_NEVER until the public macros that use it
are deprecated.

The goal here is to get internal constants for LOG_STDOUT, LOG_NEVER,
and LOG_TRACE. I want to make it clear that these log levels are custom
extensions and are not broadly supported symbols.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>