Fix: libpe_status: Add fake local node to all resources' allowed_nodes
This fixes a regression introduced by bf7ffcd. As of that commit, the
fake local node is created after all resources have been unpacked. So it
doesn't get added to resources' allowed_nodes tables.

This prevents registration of fencing devices when the fencer receives a
CIB diff that doesn't contain the local node. For example, the user may
have replaced the CIB with a boilerplate configuration that has an empty
nodes section.

See the following pull requests from Hideo Yamauchi and their
discussions:
https://github.com/ClusterLabs/pacemaker/pull/3849
https://github.com/ClusterLabs/pacemaker/pull/3852

Thanks to Hideo for the report and finding the cause.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>