Med: scheduler: Require a cluster option for new remote fencing behavior.
We don't have a ton of confidence that the previous patch is the right
thing to do for everyone, so we are going to hide it behind this
undocumented cluster config option.  By default, if the option is
missing (or is set to "true"), the existing remote fencing behavior will
be what happens.  That is, a node without quorum will be allowed to
fence remote nodes in the same partition even if they've been restarted
elsewhere.

However, with fence-remote-without-quorum="false", we will check to see
if the remote node could possibly have been started on another node and
if so, it will not be fenced.