Low: controller: Populate nodes from peer cache if can't from cluster
This code has been wrong since it was added in 2012 by 8511b3e.

I'm surprised it doesn't seem to have caused issues, and changing it now
makes me nervous because who knows what code has been built around it...
but it looks CLEARLY wrong. It may have caused little or no impact
because we used controld_node_update_quick in the places where it would
have caused a problem. Or because using the cache was basically
idempotent if we were able to load nodes from Corosync, and we had
bigger problems if we weren't.

pcmk__corosync_add_nodes() returns true if it found any nodes and added
them to node_list, and false otherwise. Until now, there have been two
possibilities:
1. pcmk__corosync_add_nodes() adds nodes to the cache and the nodes
   section (node_list), and then the caller (populate_cib_nodes()) does
   the same thing again from the cache (the "hashtable").
2. pcmk__corosync_add_nodes() fails to load any nodes. This set
   from_hashtable to FALSE, and so populate_cib_nodes() wouldn't load
   any nodes from the peer cache either.

Now, we try to load nodes from Corosync. If that fails, then we load
them from the peer cache.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>