Test: cts-fencing: Drop test for nodeid parameter inclusion
No known fence agent besides fence_dummy advertises the nodeid parameter
in its metadata, so is_nodeid_required() should never return true unless
some custom fence agent advertises nodeid and expects Pacemaker to pass
it automatically.

I checked all the agents in both fence-agents and cluster-glue.

fence_dummy advertises nodeid purely for cts-fencing's use.

So this test in cts-fencing is for a scenario that is not encountered in
practice, and it can be dropped. The nodeid parameter in fence_dummy can
be dropped accordingly.

If anyone complains about this, we can deprecate this functionality for
a while and drop it at a major or minor release. However, it seems
likely this has been unused for a long time.

It was added by 08c78ad (and again by c2265a12, oddly) in 2012. The
commit message gives no information about what motivated the change, nor
does the linked pull request give any real insight. The PR mentions
"stonith scsi" support, but the fence_scsi agent at the time did not
support a nodeid parameter.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>