Refactor: fencer: Drop stonith_device_t:priority
We've had a TODO to "hook up priority" for over 15 years: c6bac1b3. At
this point, it's safe to skip this and come back to it if needed.
Fencing levels are always an option in the meantime.

This seems like a nice-to-have at best, with fencing levels providing an
explicit way to accomplish it. (I can't think of anything priority could
do that fencing levels couldn't.)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>