Refactor: libcrmcommon: Remove max_buf_size from crm_ipc_s.
crm_ipc_s is exposed publically as crm_ipc_t, but its members are not.
Thus, we are okay to remove members from it here.

buf_size is always the same because crm_ipc_default_buffer_size() just
returns a constant, and we use qb_ipcs_enforce_buffer_size() to ensure
the IPC buffer is always 128k.  So, the block in
pcmk__connect_generic_ipc is dead code, which is the last place
max_buf_size was being used for anything.