Feature: libcrmcommon: pcmk__ipc_send_xml can send split up IPC messages.
If the IPC message is too large to fit in a single buffer, this function
now knows how to loop over the message and send it one chunk at a time.
This handles the sending side for servers that use the newer style IPC
code.  This also takes care of sending IPC events from a server.