API: libpacemaker: Document that pcmk_fence_installed() ignores timeout
It's always ignored this argument. It likely doesn't make sense to
implement a timeout, since the work currently consists of calls to
libstonith (not libstonithd!) functions and to scandir(). Implementing a
timeout would mean spawning one or more child processes and getting the
results via a pipe, which doesn't seem worth the effort.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>