Doc: libstonithd: Assume that list_agents() timeout will remain ignored
This method has always ignored the argument since it was added in 2011
via 8d6be4b6. It likely doesn't make sense to implement a timeout, since
the work currently consists of calls to libstonith (not libstonithd!)
functions and to scandir(). Implementing a timeout would mean spawning
one or more child processes and getting the results via a pipe, which
doesn't seem worth the effort.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>