Test: cts-cli: Make crm_diff test pass again
Commit bb20fd93 introduced a regression. However, there were about four
weeks between the last time CI tests were run against the corresponding
PR and when the PR was merged. In the interim, a cts-cli test for
crm_diff was added. (In particular, it was re-added after a mistaken
deletion).

It turns out commit bb20fd93 breaks that test.

This should be wholly insignificant in practice. The result is that a
comment that was unchanged may appear in a diff as a deletion followed
by a creation, if preceding sibling nodes were deleted or moved. The
pcmk__xf_skip flag is no longer being added to deleted nodes before we
check whether a comment's old position matches its new position.

This is looking like a pain to fix, and I want to ensure that other PRs
don't fail CI due to this issue. That is the purpose of this commit.

As an aside, I don't think we should support diffing XML comments, since
Pacemaker doesn't care about them. This support complicates our XML diff
code CONSIDERABLY.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>