Refactor: libcrmcommon: Drop xmlSetBufferAllocationScheme() call
As of libxml v2.14.0, xmlSetBufferAllocationScheme() is deprecated,
which is already causing compilation failures on debian-experimental.
Further, also as of libxml v2.14.0, its replacement
xmlBufferSetAllocationScheme() is a no-op.

At this point, we could easily drop pcmk__xml_init(), replacing it with
calls to pcmk__schema_init(). However, pcmk__xml_cleanup() still seems
reasonable to have, and it seems intuitive to keep a corresponding
init() function. So for now this remains as a wrapper.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>