API: libcrmcommon: Introduce the pcmk_rc_ipc_more error code.
Functions should return this error code if a single IPC message is too
large to fit into a single buffer, and so it has to be split up into
multiple.  It can also be used when one chunk of a split up message has
been read, indicating that the caller needs to continue reading.