Refactor: libcrmcommon: EAGAIN isn't an error for pcmk__ipc_send_iov...
...at least, when it's sending a server event.  In this case, the iov
will be added to the send queue and EAGAIN will be returned.  We will
attempt to send the event again the next time pcmk__ipc_send_iov is
called, or when the timer pops to flush the event queue.

Thus, we are already handling the EAGAIN return code properly elsewhere
and just need to proceed as if everything will be fine.