Refactor: libcrmcommon: discard_old_replies can handle multipart messages.
This function now reads a single IPC message.  As long as we haven't
read a standalone message or the last part of a multipart message, we
just keep blocking sending ensuring this function will get called again.

Additionally, there should be no problems if we get EAGAIN when trying
to read the message.  In this case, we'll just keep sending blocked and
try again next time we try to send.