Refactor: libcrmcommon: Remove buf_size from crm_ipc_s.
The libqb read buffer is now a fixed size, so we can just use
crm_ipc_default_buffer_size() anywhere we need that.  And then anywhere
we need to know the size of buffer (nowhere at the moment, but possibly
in the future) we can just use GByteArray->len to figure that out.  So
there's no need to keep an extra struct member around anymore.