Low: libcrmcommon: All parts of a multipart event should have the same ID
Events that aren't a response to a request start off with an ID of 0.
Later in pcmk__ipc_send_iov, we give them an ever-increasing ID number.
Because multipart events are processed by just looping through the
message and sending parts of it, each part will get its own unique ID
number.

However, on the receive side, we check that each part of a multipart IPC
event has the same ID as the first part and error out if this is not the
case.  Therefore, the send side needs to make sure this happens.