Feature: libcrmcommon: crm_ipc_send can send large IPC messages.
Previous patches focused on crm_ipc_send accepting multi-part IPC
replies, and then on the server side sending those replies.  This patch
adds code to crm_ipc_send to allow it to send multi-part IPC messages to
begin with.

EAGAIN is a problem here as well, since we don't have a send queue for
regular IPC messages, but at least there's a timeout.  If we get EAGAIN
on some intermediate part of the message, it's going to cause problems
since the server will only have received part of it.  The server will
still be able to clean up after the incomplete message, at least.