Low: Address certain clang static analysis errors.
There are still a bunch remaining.  clang seems to have trouble with not
understanding that we are checking error conditions through our own
assertion and check macros.  It also has problems recognizing that out
can't be NULL if pcmk__output_new returns pcmk_rc_ok, but only
sometimes.  In general, it seems to only pick up on some of these errors
sometimes.

The fixes for those issues don't seem to add any value to the code, but
they do make it look more complicated.  Thus I'm not going to fix those
at the moment.  The things in this patch are more reasonable fixes.