Low: libs: Fix cppcheck warnings.
A couple of these are just suppressions because cppcheck doesn't
understand some assertion or previous condition that prevents a pointer
from being NULL.  A few of these are actual problems, however.

Note that there are still a lot of cppcheck warnings, but the remaining
ones all come from it thinking that g_string_sized_new could return
NULL.  However, that function uses g_slice_new internally, and the
documentation for that function claims it can never return NULL.
cppcheck doesn't understand this, and there's not a good way to turn off
the warning for that short of adding a suppression message at every
caller.