Low: Address coverity issues.
For the most part, this is just introducing comments to tell it to not
worry about things, though there were a couple uninitialized variables
and one minor memory leak to deal with.

Some of the things I'm telling it to ignore are problems that aren't
worth fixing because they're not causing problems and have been removed
from the main branch (like the XML schema code).  Others can't be fixed
in a minor release (we can't start sanitizing environment variables, for
instance).