Refactor: cts: Remove --valgrind-tests
This option was meant to enable a special ValgrindTest class in cts-lab.
However, when I rearranged all the cts-lab code, I didn't bring that
test class along.  This is probably because in order for it to do
anything, you would have to enable valgrind on each lab node, but it
still wouldn't do anything because the test's __call__ method was empty.

I also find this option's presence confusing.  I expected it to actually
enable valgrind for me, which it did not do.  So just remove it.

Related T989