Doc: Pacemaker Explained: Clarify alert timestamp-format
A leading zero is ignored for "%xN", so remove it from the example. Add
detail to the description.

Also, we strftime() into a fixed-size buffer of 128 characters. For some
reason, there is no general way to strftime into a dynamically sized
buffer or to determine how large the buffer needs to be.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>