Refactor: executor: Some best practices in remoted_spawn_pidone()
The reasoning behind a lot of this code is unclear to me. For example,
why is HAVE_PROGNAME relevant when we never use __progname or
__progname_full? Why do we overwrite the envp members, and why do we
assume they might be located immediately after the argv array?

This commit just implements some very basic best practices and standard
naming to a block within remoted_spawn_pidone(). We can revisit later.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>