Revert "Merge pull request #3885 from clumens/ipc-no-compression-2.1"
This reverts commit 8da457d872661c794e9191e9306c486bd3ef2c83, reversing
changes made to d775c79f65e81b9a6513ff7f94f2a87a33475420.

NOTE:  This PR shouldn't have been merged to the 2.1 branch.  It's the
first part of three IPC split up PRs, and one of the things it does is
disable compression for IPC messages.  This will break installations for
anyone with an even moderately large CIB.  Basically, this PR can only
be merged if the other two parts are going to be, and we're not ready to
do that yet.