Refactor: libcrmcommon: Avoid snprintf() in crm_log_filter()
To avoid the need to error-check snprintf() explicitly, which we have
not been doing. This also avoids imposing an arbitrary (albeit
reasonable) limit of 500 characters on tag length.

It should be simple to replace some of this logic with a call to
g_strsplit(). However, that would require testing that I don't feel like
doing right now. The current change seems straightforward enough.

Ref T974

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>