Fix: libcrmcommon: Increase poll() timeout to 5s for liveness checks on sub-daemons
If a node is temporarily running into a difficult situation, for the
liveness checks on sub-daemons by pacemakerd, the 5
`PCMK_PROCESS_CHECK_RETRIES` offer a certain degree of tolerance. But in
practice with the current 2s of poll() timeout, if the situation is a
little too difficult, the liveness checks seem to run into 5 failures in
a row rather easily while the situation still seems acceptable.

This commit offers a longer timeout of 5s for a sub-daemon to respond to
a liveness check.

See discussion at:
https://github.com/ClusterLabs/pacemaker/pull/2588#discussion_r2122975853