API: schemas: Add a schema for crm_diff
This is pretty basic. We can have either a "patchset" element or an
"updated" element, and each just contains a CDATA block.

It would be possible to use the diff API schema and include the patchset
XML directly rather than as CDATA. This is less straightforward to do
for the "updated" XML. The input is expected to be a CIB, so I suppose
we could use the CIB schema there.

Unfortunately the patchset itself already uses "diff" as its root
element, so I wanted to avoid using "diff" as a top-level element name
for crm_diff output. It also wouldn't make a lot of sense when applying
a patchset (that is, where we're displaying an "updated" element).

Closes T114

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>