Test: libcrmcommon: Use June for pcmk__time_format_hr() test date
Currently ha_get_tm_time() assigns the source crm_time_t object's years
(day of year) member to the mday (day of month) member of the target
struct tm.

One would expect that we would instead assign it to the target's yday
(day of year) member.

However, the call to mktime() fixes out-of-range values to make all the
target's fields consistent if possible. For example, October 40 becomes
November 9. The original yday value is ignored. This is why we assign
the day of the year to the mday member.

If there were any problem with this logic (for example, if mktime() were
omitted), then a test date in January may still give correct results.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>