Test: libcrmcommon: Test characters after %N
in pcmk__time_format_hr() unit tests. The parsing logic in
pcmk__time_format_hr() is hard to follow. For a little while, I was
confused and thought it would give incorrect output if there were any
format specifiers after a "%N".

There is probably no real use case for this support, but it should work
(and currently it does work).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>